requirements.txt
asgiref==3.8.1
certifi==2024.2.2
Django==5.0.3
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
